fallen_deliver/requirements.txt
pyopenssl
ndg-httpsclient
pyasn1
shyaml==0.6.1
pysnooper==0.4.3
gitpython==3.1.14
chardet==3.0.4
idna==2.9
certifi==2020.4.5.1
urllib3==1.25.9
requests==2.23.0
pytz==2020.1
tzlocal==2.0.0
six==1.14.0
PyYAML==5.3.1
click==7.1.2
itsdangerous==1.1.0
MarkupSafe==1.1.1
requests-toolbelt==0.9.1
SQLAlchemy==1.4.22
PyMySQL==1.0.2
loguru==0.5.3
jinja2==2.11.3
wechat_work_webhook==0.0.2
python-redis-lock==3.7.0
jsonpath==0.82